MMM adjustments
Increased the fTP of "Cursed Sphere Nightmare" used by Terrorfly to 5(was 1.5)

Increased the amount of status effects drained by Debilitating Drone to 7(was 1 random)

Increased the fTP of "Somersault" to 2.0(was 1.5)

Terror Flys Amnesia aura should now have 0 downtime in it's application.

Awful Eye Nightmare now properly resets hate when the enemy is facing Tanihwa.